Conflicts:
	CornerStone.pri
	MultiTouch/BGCalibrator.cpp
	MultiWidgets/Calibration.cpp
	Plugins/LightBeings/Snake.cpp
	Plugins/Plugins.pri
	Scripting/JS/Console.cpp
	Scripting/generator/exclude_files
	multitude/Luminous/Image.cpp
	multitude/Nimble/Vector2.hpp
	multitude/Nimble/Vector3.hpp
	multitude/Radiant/Platform.hpp
	multitude/Radiant/Radiant.pro
	multitude/library.pri
	multitude/multitude.pri
	multitude/multitude.pro

+++ multitude/Radiant/SMRingBuffer.cpp
@@ -511,7 +511,7 @@ namespace Radiant
   {
     int bytes = data.pos();
     const void * ptrs[2] = { & bytes, data.data() };
-    uint32_t sizes[2] = { 4, data.pos() };
+    uint32_t sizes[2] = { 4, uint32_t(data.pos()) };
 
     return write(2, ptrs, sizes);
   }
